refactor(#3631): split the marshal TYPE-STRING parser out of udt.rs
Pure move, no behaviour change: `parse_udt_type_definition`,
`parse_cassandra_type_with_depth` and their three string helpers go to
`udt/type_string.rs`, whose header records the pinned-tag authority for the
marshal-name -> CQL-type mapping (`cassandra-5.0.8` CQL3Type.Native).

udt.rs 1401 -> 1124 lines (campsite rule, epic #1116); the new home is where
roborev job 68's finding 1 fix lands.
